+++ src/main/java/org/astrojournal/generator/absgen/LatexExporterByConst.java
@@ -47,7 +47,6 @@ import org.astrojournal.generator.Report;
  * @author Piero Dalle Pezze
  * @version $Rev$
  * @since 1.0
- * @date 17 Jan 2016
  */
 public abstract class LatexExporterByConst extends LatexExporter {