+++ src/main/java/org/astrojournal/generator/extgen/ExtLatexExporterByDate.java
@@ -37,8 +37,8 @@ import org.astrojournal.generator.absgen.LatexExporterByDate;
  * export.
  * 
  * @author Piero Dalle Pezze
- * @version 0.3
- * @since 28/05/2015
+ * @version $Rev$
+ * @since 1.0
  */
 public class ExtLatexExporterByDate extends LatexExporterByDate {