+++ src/main/java/org/astrojournal/gui/menu/AJMenuBar.java
@@ -51,7 +51,6 @@ import org.astrojournal.gui.dialogs.LicenseDialog;
  * @author Piero Dalle Pezze
  * @version $Rev$
  * @since 1.0
- * @date 13 Dec 2015
  */
 public class AJMenuBar extends JMenuBar implements ActionListener {
 
@@ -67,6 +66,7 @@ public class AJMenuBar extends JMenuBar implements ActionListener {
 
     private JMenu fileMenu;
     private JMenuItem fileCreateJournal;
+    private JMenuItem fileOpenJournal;
     private JMenuItem fileQuit;
 
     private JMenu editMenu;
@@ -109,7 +109,19 @@ public class AJMenuBar extends JMenuBar implements ActionListener {
 	fileCreateJournal.addActionListener(this);
 	fileMenu.add(fileCreateJournal);
 
-	// fileMenu.addSeparator();
+	fileOpenJournal = new JMenuItem(
+		resourceBundle.getString("AJ.cmdOpenJournal.text"));
+	fileOpenJournal.setIcon(new ImageIcon(ClassLoader
+		.getSystemResource("graphics/icons/open_journals_16.png")));
+	fileOpenJournal.setMnemonic(KeyEvent.VK_O);
+	fileOpenJournal.setAccelerator(KeyStroke.getKeyStroke('O', Toolkit
+		.getDefaultToolkit().getMenuShortcutKeyMask()));
+	fileOpenJournal.setActionCommand(AJGUIActions.OPEN_JOURNAL.name());
+	fileOpenJournal.addActionListener(this);
+	fileOpenJournal.setEnabled(false);
+	fileMenu.add(fileOpenJournal);
+
+	fileMenu.addSeparator();
 
 	fileQuit = new JMenuItem(resourceBundle.getString("AJ.mnuQuit.text"));
 	fileQuit.setIcon(new ImageIcon(ClassLoader
@@ -176,6 +188,9 @@ public class AJMenuBar extends JMenuBar implements ActionListener {
 	if (action.equals(AJGUIActions.CREATE_JOURNAL.name())) {
 	    application.createJournals();
 
+	} else if (action.equals(AJGUIActions.OPEN_JOURNAL.name())) {
+	    application.openJournals();
+
 	} else if (action.equals(AJGUIActions.QUIT.name())) {
 	    application.quit();
 
@@ -223,6 +238,8 @@ public class AJMenuBar extends JMenuBar implements ActionListener {
     public void setEnabled(String action, boolean enabled) {
 	if (action.equals(AJGUIActions.CREATE_JOURNAL.name())) {
 	    fileCreateJournal.setEnabled(enabled);
+	} else if (action.equals(AJGUIActions.OPEN_JOURNAL.name())) {
+	    fileOpenJournal.setEnabled(enabled);
 	} else if (action.equals(AJGUIActions.EDIT_PREFERENCES.name())) {
 	    editPreferences.setEnabled(enabled);
 	} else {