+++ src/main/java/org/astrojournal/utilities/PropertiesManager.java
@@ -36,7 +36,6 @@ import java.util.Properties;
  * @author Piero Dalle Pezze
  * @version $Rev$
  * @since 1.0
- * @date 18 Jan 2016
  */
 public class PropertiesManager {