+++ src/main/java/org/astrojournal/utilities/RunExternalCommand.java
@@ -35,7 +35,6 @@ import java.util.ResourceBundle;
  * @author Piero Dalle Pezze
  * @version $Rev$
  * @since 1.0
- * @date 22 Dec 2015
  */
 public class RunExternalCommand {