+++ src/test/java/main/AJConfigurationTestUtils.java
@@ -31,7 +31,6 @@ import org.astrojournal.configuration.ajconfiguration.AJPropertyConstants;
  * @author Piero Dalle Pezze
  * @version $Rev$
  * @since 1.0
- * @date 15 Jan 2016
  */
 public class AJConfigurationTestUtils {