+++ src/test/java/main/MiniTSVImporterTest.java
@@ -50,7 +50,6 @@ import org.springframework.context.support.ClassPathXmlApplicationContext;
  * @author Piero Dalle Pezze
  * @version $Rev$
  * @since 1.0
- * @date 15 Jan 2016
  */
 public class MiniTSVImporterTest {